Inline some utility methods
